+++ R/post-proc.R
@@ -1,18 +1,4 @@
 #' @export
-#' @title add images into an rdocx object
-#' @description reference images into a Word document.
-#' This function is now useless as the processing of images
-#' is automated when using [print.rdocx()].
-#'
-#' @param x an rdocx object
-#' @param src a vector of character containing image filenames.
-#' @family functions for officer extensions
-#' @keywords internal
-docx_reference_img <- function(x, src) {
-  x
-}
-
-#' @export
 #' @title transform an xml string with images references
 #' @description This function is useless now, as the processing of images
 #' is automated when using [print.rdocx()].